stcomeçando a descompactar

+++ caractere.c
@@ -5,23 +5,23 @@
 struct caractere
 {
     unsigned char symbol;
-    int frequence;
+    long int frequence;
 };
 
-Caractere* generateCh(unsigned char symbol, int frequence){
+Caractere* generateCh(unsigned char symbol, long int frequence){
     Caractere* ch = (Caractere*)malloc(sizeof(Caractere));
     ch->symbol = symbol;
     ch->frequence = frequence;
     return ch;
 }
 
-int returnFrequence(Caractere *c){
+long int returnFrequence(Caractere *c){
     return c->frequence;
 }
 unsigned char returnSymbol(Caractere *c){
     return c->symbol;
 }
 void printCharactere(Caractere *c){
-    printf("(%c|%d)", c->symbol, c->frequence);
+    printf("(%c|%ld)", c->symbol, c->frequence);
 }
